requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
keras>=2.13.0

# Gradient Boosting
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Feature Engineering
featuretools>=1.28.0
tsfresh>=0.20.0

# Graph Analysis
networkx>=3.1
torch-geometric>=2.4.0
dgl>=2.0.0

# Privacy-Preserving ML
flwr>=1.6.0  # Federated Learning
opacus>=1.4.0  # Differential Privacy
tenseal>=0.3.10  # Homomorphic Encryption

# Streaming
kafka-python>=2.0.2
faust-streaming>=0.10.0
redis>=5.0.0

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# MLOps & Monitoring
mlflow>=2.9.0
evidently>=0.4.0
prometheus-client>=0.19.0
grafana-api>=1.0.0

# Explainability
shap>=0.43.0
lime>=0.2.0.1
captum>=0.6.0  # PyTorch interpretability

# Data Validation
great-expectations>=0.18.0
pydantic>=2.5.0

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
locust>=2.20.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0
joblib>=1.3.0
pickle5>=0.0.11

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pymongo>=4.6.0

# Cloud & Container
docker>=7.0.0
kubernetes>=28.1.0
boto3>=1.34.0

# Logging & Debugging
loguru>=0.7.0
pdbpp>=0.10.3
